BUGFIX Removed restrictive decimal for Tax Rate field and replaced with Double. Ticket #2976
git-svn-id: http://svn.silverstripe.com/open/modules/ecommerce/trunk@74572 467b73ca-7a2a-4603-9d3b-597d59a354a9

+++ code/modifiers/TaxModifier.php
@@ -17,7 +17,7 @@ class TaxModifier extends OrderModifier {
 	
 	static $db = array(
 		'Country' => 'Text',
-		'Rate' => 'Decimal',
+		'Rate' => 'Double',
 		'Name' => 'Text',
 		'TaxType' => "Enum('Exclusive,Inclusive')"
 	);
